requirements.txt
flask~=1.1.2
requests~=2.25.1
SQLAlchemy~=1.3.24
Werkzeug~=1.0.1
flask_login
flask_sqlalchemy
flask_migrate
flask_restful
numpy
Pillow~=8.3.2
